feat: added configurable updateStrategy to server deployment
This enables setting the spec.strategy to Recreate on the server
deployment. This includes checks in the checks.yaml to have some nice
UX.
Without this, having a storage backend that does not allow for a
multiattached pv the current chart always results in manual intervention
(manual deletion of the old pod or restarting the entire deployment).